Fixed old sdk compilation issues
git-svn-id: http://miranda.googlecode.com/svn/trunk@9190 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/utils/path.cpp
@@ -395,7 +395,7 @@ XCHAR *GetInternalVariable(XCHAR *key, int keyLength, HANDLE hContact)
 }
 
 template<typename XCHAR>
-XCHAR *GetVariableFromArray(REPLACEVARSARRAY *vars, XCHAR *key, int keyLength, HANDLE hContact, bool *bFree)
+XCHAR *GetVariableFromArray(REPLACEVARSARRAY *vars, XCHAR *key, size_t keyLength, HANDLE hContact, bool *bFree)
 {
 	*bFree = false;
 	for (REPLACEVARSARRAY *var = vars; var && var->lptzKey; ++var)
